query-simulator/requirements.txt
prometheus-client
psycopg2-binary